requirements.txt
langchain>=0.1.0
langchain_openai>=0.2.0
python-docx>=0.8.11
pytesseract>=0.3.10
easyocr>=1.7.1
pydantic>=2.5.2
python-multipart>=0.0.6
streamlit>=1.30.0
openai>=1.3.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0 